get rid of core K&R, use gnulib's _GL_UNUSED and avoid some signed/unsigned comparisons
This significantly reduces the number of warnings if requesting a
slightly bigger set of warnings.

+++ gv/src/secscanf.c
@@ -282,7 +282,7 @@ sec_sscanf(const char *s, const char *fmt, ...)
 					goto stop;
 				buf = va_arg(ap, char *);
 				size = va_arg(ap, size_t) - nul_terminated;
-				if (size > s - string)
+				if (size > (size_t)(s - string))
 					size = s - string;
 				strncpy(buf, string, size);
 				if (nul_terminated)
